// packages/typespec-test/test/healthInsights_radiologyinsights/generated/typespec-ts/src/api/radiologyInsights/options.ts
// Copyright (c) Microsoft Corporation.
// Licensed under the MIT License.

import { RadiologyInsightsModelConfiguration } from "../../models/models.js";
import { OperationOptions } from "@azure-rest/core-client";

/** Optional parameters. */
export interface RadiologyInsightsCreateJobOptionalParams extends OperationOptions {
  /** Delay to wait until next poll, in milliseconds. */
  updateIntervalInMs?: number;
  /** Configuration affecting the Radiology Insights model's inference. */
  configuration?: RadiologyInsightsModelConfiguration;
  /** An opaque, globally-unique, client-generated string identifier for the request. */
  repeatabilityRequestId?: string;
  /** Specifies the date and time at which the request was first created. */
  repeatabilityFirstSent?: Date;
}

/** Optional parameters. */
export interface RadiologyInsightsGetJobOptionalParams extends OperationOptions {}